+++ Common/Source/Draw/LKProcess.cpp
@@ -1883,7 +1883,7 @@ goto_bearing:
 		// B88 B89
 		case LK_EXP1:
 			_stprintf(BufferTitle, TEXT("RAM"));
-			_stprintf(BufferValue, TEXT("%zu"),(CheckFreeRam()/1024));
+			_stprintf(BufferValue, TEXT("%u"),(unsigned int)(CheckFreeRam()/1024));
                         valid=true;
                         break;
